Refactor login ID config
