+++ pkg/auth/dependency/userverify/senderfactory.go
@@ -40,7 +40,7 @@ func (d *defaultCodeSenderFactory) NewCodeSender(urlPrefix *url.URL, loginIDKey 
 	if !ok {
 		panic("invalid user verification login id key: " + loginIDKey)
 	}
-	authLoginIDKey, ok := d.Config.AppConfig.Auth.GetLoginIDKey(loginIDKey)
+	authLoginIDKey, ok := d.Config.AppConfig.Identity.LoginID.GetKey(loginIDKey)
 	if !ok {
 		panic("invalid login id key: " + loginIDKey)
 	}
